fix(ai-agents): keep aiGuardrails.ts free of registry/schema imports
toolActionEnum moved to aiToolActions.ts. The review refactor had it in
aiGuardrails.ts importing aiToolSchemas (which imports Drizzle enums from
db/schema), which broke six Test API suites that partially mock db/schema and
transitively import guardrails. A contract test now pins the import surface.
